// src/core/enums/action-log.enum.ts
export enum ActionLog {
  SIGN_UP = 'sign-up',
  SIGN_IN = 'sign-in',

  NPC_START_INTERACTION = 'npc-start-interaction',
  NPC_END_INTERACTION = 'npc-end-interaction',
  NPC_SENT_MESSAGE = 'npc-sent-message',

  SLIDESHOW_START = 'slideshow-start',
  SLIDESHOW_END = 'slideshow-end',
  SLIDESHOW_SEEN_SLIDE = 'slideshow-seen-slide',

  QUIZ_START = 'quiz-start',
  QUIZ_END = 'quiz-end',
  QUIZ_ANSWER = 'quiz-answer',
}